Adds ability to promote/banish subtopics
In a discussion, the promote and banish methods accept a comment,
then finds all child comments of it. It then creates a new discussion
with a summary matching the comment that was passed, and moves all
child comments into it.

The difference between promote and banish is that banish sets the
discussion's visibility to false (which does nothing now, but will
mean that it's not advertised on its source page, and its comments are
not checked for further promotion/banishment)