Refines negativity checking
First, reduces default threshold for articles to 5 so it's easier to
trip the detection.

Second, after each comment update, the attempt to detect a negative
thread only happens if the commen has a negative rating already. Also,
the check to find the oldest ancestor in a negatively rated thread
happens first, then the pruning continues as it did previously.